+++ src/components/navigation/navigation.tsx
@@ -36,8 +36,6 @@ function Navigation(): JSX.Element {
             <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
               <div className="header__avatar-wrapper user__avatar-wrapper">
               </div>
-              <span className="header__user-name user__name">Привет, гость!</span>
-              <span style={{ margin: '10px' }}></span>
               <span className="header__login">Sign in</span>
             </Link>
           </li>
